Separate utils functions to utils/
